bug fix cloudinary cloud name

+++ src/utils/cloudnary.js
@@ -1,17 +1,21 @@
 import { v2 as cloudinary } from "cloudinary"
-import { error } from "console"
 import fs from "fs"
+
+
 
 // basic config 
 cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_SECRET_KEY,
+    cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
+    api_key:process.env.CLOUDINARY_API_KEY,
+    api_secret:process.env.CLOUDINARY_SECRET_KEY,
     secure: true
 })
 
+
+
 // create function to handle file upload
 const uploadOnCloudinary = async (localFilePath) => {
+   
     try {
         if (!localFilePath) { throw "File path is required" }
 
@@ -22,13 +26,23 @@ const uploadOnCloudinary = async (localFilePath) => {
 
         // file has been uplaoded successfully
         console.log("file is uploaded successfully")
+             // delete temp file after successful upload
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
 
         return response
 
     } catch (error) {
 
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null
+       // delete temp file if upload failed
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        console.error("Cloudinary upload error:", error);
+
+        return null;
     }
 }
